Refactor: 배경이미지, 티켓 정보 저장 userId 로직 변경
배경이미지, 티켓 정보 저장 userId 로직 변경
- 로그인된 사용자는 JWT를 통해 userId를 알아내고, 비로그인 사용자는 null로 처리해서 DB 테이블에 저장
- Category, Tag, TagBackgroundImage 엔티티 생성(다음 작업 예정)

+++ src/main/java/org/chunsik/pq/model/User.java
@@ -14,7 +14,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String nickname;
@@ -31,7 +31,7 @@ public class User {
     @Enumerated(EnumType.STRING)
     private OauthProvider oauthProvider;
 
-    private User(Long id, String nickname, String email, String password, LocalDateTime createdAt, OauthProvider oauthProvider) {
+    private User(Integer id, String nickname, String email, String password, LocalDateTime createdAt, OauthProvider oauthProvider) {
         this.id = id;
         this.nickname = nickname;
         this.email = email;
